Performance improvement by avoiding rendering the entire output buffer as Handlebars
Instead, this commit changes the perform_merge function to render the JSON to be
merged with the output buffer, which is typically going to be smaller than our
log message, and then merges that result.

The underlying assumption here is that Handlebars is going to take longer on
longer strings, and is costing more time than an extra serialization with serde.